+++ mapiproxy/servers/default/emsmdb/oxctabl.c
@@ -771,7 +771,7 @@ _PUBLIC_ enum MAPISTATUS EcDoRpc_RopFindRow(TALLOC_CTX *mem_ctx,
 
 	memset (&row, 0, sizeof(DATA_BLOB));
 
-	switch (emsmdbp_is_mapistore(object)) {
+	switch ((int)emsmdbp_is_mapistore(object)) {
 	case true:
 		/* Restrict rows to be fetched */
 		retval = mapistore_table_set_restrictions(emsmdbp_ctx->mstore_ctx, emsmdbp_get_contextID(object), object->backend_object, &request.res, &status);
