+++ mapiproxy/servers/default/nspi/emsabp.c
@@ -814,7 +814,7 @@ _PUBLIC_ enum MAPISTATUS emsabp_table_fetch_attrs(TALLOC_CTX *mem_ctx, struct em
 				emsabp_PermanentEntryID_to_Binary_r(mem_ctx, permEntryID, &(lpProps.value.bin));
 				break;
 			case PR_CONTAINER_FLAGS:
-				switch (child) {
+			  switch ((int)child) {
 				case true:
 					lpProps.value.l = AB_RECIPIENTS | AB_SUBCONTAINERS | AB_UNMODIFIABLE;
 					break;
@@ -823,7 +823,7 @@ _PUBLIC_ enum MAPISTATUS emsabp_table_fetch_attrs(TALLOC_CTX *mem_ctx, struct em
 				}
 				break;
 			case PR_DEPTH:
-				switch (child) {
+			  switch ((int)child) {
 				case true:
 					lpProps.value.l = 0x1;
 					break;
